Refactor & add documentation
