+++ src/middleware/metrics/index.ts
@@ -1,3 +1,2 @@
 export * from './counters'
-export * from './responseMetadata'
 export * from './responseProfiler'
