+++ src/middleware/standardResponses/index.ts
@@ -1,3 +1,4 @@
+export * from './getResponseMetadata'
 export * from './jsonApi'
 export * from './standardErrors'
 export * from './standardResponses'
